Backend/requirements.txt
Flask==3.0.2
Flask-Cors==4.0.0
Flask-JWT-Extended==4.6.0
Flask-Migrate==4.0.5
SQLAlchemy==2.0.23
alembic==1.12.1
razorpay==1.4.2
python-dotenv==1.0.0
WeasyPrint==59.0
PyJWT==2.8.0
Werkzeug==3.0.1
requests==2.31.0
pytest==7.4.4
pytest-flask==1.3.0
reportlab==4.0.6

